+++ app/controllers/skill_managements_controller.rb
@@ -1,9 +1,12 @@
 class SkillManagementsController < ApplicationController
+  before_action :set_personal_skill, only: %w[edit show]
+
   def show
-    @skill_management = SkillManagement.where(user_id:current_user.id)
   end
 
   def edit
+    @skill_managements = SkillManagement.find(id: params[:id])
+
   end
 
   def new
@@ -13,20 +16,25 @@ class SkillManagementsController < ApplicationController
   def create
     @form = Form::SkillManagementCollection.new(skill_management_collection_params)
     if @form.save
-      redirect_to skill_managements_path, notice: "skillを登録しました"
+      redirect_to skill_managements_path, notice: 'skillを登録しました'
     else
-      flash.now[:alert] = "skill登録に失敗しました"
+      flash.now[:alert] = 'skill登録に失敗しました'
       render :new
     end
+  end
 
+  def update
+    @skill_managements.update(skill_management_collection_params)
   end
 
   private
 
-    def skill_management_collection_params
-        params.require(:form_skill_management_collection)
-        .permit(skill_managements_attributes: [:user_id, :skill_id, :skill_level, :availability])
-    end
+  def skill_management_collection_params
+    params.require(:form_skill_management_collection)
+          .permit(skill_managements_attributes: [:user_id, :skill_id, :skill_level, :availability])
+  end
 
-
+  def set_personal_skill
+    @personal_skill = SkillManagement.where(user_id: current_user.id)
+  end
 end
