+++ app/models/form/Base.rb
@@ -1,7 +1,6 @@
 class Form::Base
-    include ActiveModel::Model
-    include ActiveModel::Callbacks
-    include ActiveModel::Validations
-    include ActiveModel::Validations::Callbacks
-  
-  end
+  include ActiveModel::Model
+  include ActiveModel::Callbacks
+  include ActiveModel::Validations
+  include ActiveModel::Validations::Callbacks
+end
