+++ app/models/skill_management.rb
@@ -1,6 +1,6 @@
 class SkillManagement < ApplicationRecord
-belongs_to :user
-belongs_to :skill, optional: true
+  belongs_to :user
+  belongs_to :skill, optional: true
 
-enum skill_level: { Junior: 0, Middle: 1, Senior: 2 }
+  enum skill_level: { Junior: 0, Middle: 1, Senior: 2 }
 end
