+++ app/views/managements/index.html.erb
@@ -11,7 +11,7 @@
   <td>稼働時間</td>
   <td>単価</td>
   <td>単位</td>
-  <td>アクション</tg>
+  <td>アクション</td>
 </tr>
 
 <% @management.each do |m| %>
