+++ app/views/mypages/show.html.erb
@@ -16,7 +16,11 @@
         <li>メールアドレス: <%= @user.email %></li>
         <li>氏名:          <%= @user.name %></li>
         <li>電話番号:       <%= @user.phone_number %></li>
-        <li>スキル:         <%= @user.skills %></li>
+        <li>スキル:         <% @user.skills.each do |s| %>
+            <%= s.skill %>
+            <%= s.skill_level %>
+            <% end %>
+            </li>
         <li>スキルレベル:    <%= @user.skill_level %></li>
         <li>Githubアカウント: <%= @user.github_url %></li>
         <li>口座情報:       <%= @user.account_info %></li>
