+++ app/views/skill_managements/edit.html.erb
@@ -1,3 +1,33 @@
 <h1>SkillManagements#edit</h1>
 <p>Find me in app/views/skill_managements/edit.html.erb</p>
 
+<h2>スキル一括登録 </h2>
+<%= form_with model: @form, url: skill_managements_path, method: :patch, id: 'new_skill_management', class: 'new_skill_management', local: true do |form| %>
+    <table class="table">
+    <thead>
+     <tr>
+        <th> 登録
+        <th> スキル
+        <th> スキルレベル
+     </tr>
+    </thead>
+  <tbody>
+    <%= form.fields_for :skill_managements do |f| %>
+      <tr>
+        <td class= "text-center">
+        <%= f.check_box :availability %>
+        </td>
+        <td>
+        <%= f.collection_select :skill_id, Skill.all, :id, :skill, :prompt => "スキルを選択してください" %><br>
+        </td>
+        <td>
+        <%= f.select :skill_level, Skill.skill_levels.keys.map {|k| [k, k]}, {class: 'form-control', required: true } %>
+        </td>
+        <%= f.hidden_field :user_id, :value => current_user.id %>
+    <% end %>
+    </table>
+
+    <%= form.submit "一括登録" %>
+    <%= link_to "戻る", :back  %>  
+
+<% end %>
