+++ app/views/skill_managements/show.html.erb
@@ -5,19 +5,23 @@
 
 <%= link_to "新規登録", new_skill_managements_path %><br>
 
+
+
 <table class="">
 <tbody>
 <tr>
   <td>スキル</td>
   <td>スキルレベル</td>
-
+  <td></td>
 </tr>
 
 
-<% @skill_management.each do |s| %>
+<% @personal_skill.each do |s| %>
     <tr>
     <td><%= s.skill.skill %></td>
     <td><%= s.skill_level %></td>
+    <td> <%= link_to "編集", edit_skill_managements_path(id: s.id) %></td>
+
     </tr>
 <% end %>
 </tbody>
