+++ test/controllers/skill_managements_controller_test.rb
@@ -1,12 +1,12 @@
-require "test_helper"
+require 'test_helper'
 
 class SkillManagementsControllerTest < ActionDispatch::IntegrationTest
-  test "should get show" do
+  test 'should get show' do
     get skill_managements_show_url
     assert_response :success
   end
 
-  test "should get edit" do
+  test 'should get edit' do
     get skill_managements_edit_url
     assert_response :success
   end
