+++ test/models/skill_management_test.rb
@@ -1,4 +1,4 @@
-require "test_helper"
+require 'test_helper'
 
 class SkillManagementTest < ActiveSupport::TestCase
   # test "the truth" do
